add autentikasi, session, and cookie

+++ main/views.py
@@ -5,9 +5,15 @@ from django.core import serializers
 from django.http import HttpResponse
 from django.shortcuts import get_object_or_404, redirect, render
 from main.forms import SkillForm, EducationForm, ExperienceForm
-
+from django.contrib.auth import login, logout
+from django.contrib.auth.forms import AuthenticationForm, UserCreationForm
+from django.shortcuts import redirect, render
+import datetime
+from django.contrib.auth.decorators import login_required 
+from django.core.exceptions import PermissionDenied        
 
 def show_main(request):
+    last_login = request.COOKIES.get('last login', 'Belum ada sesi login / Cookie tidak ditemukan')
     context = {
         "name": "Salma Maharani",
         "npm": "2506586532",
@@ -19,8 +25,8 @@ def show_main(request):
             "Big on trying new hobbies even if I abandon them two weeks later, but hey, we love a versatile queen. "
             "Listening to music like it's a full-time job, and trying not to let my tech stack or my sleep schedule break down completely. "
             "Pls send cat pics immediately.🐾🐾🐾🐾🐾🐾🐾🐾🐾🐾🐾🐾🐾"
-
         ),
+        "last_login": last_login,
     }
     return render(request, "index.html", context)
 
@@ -32,7 +38,10 @@ def show_experience(request):
     }
     return render(request, "experience.html", context)
 
+@login_required(login_url="/login/") 
 def create_experience(request):
+    if not request.user.is_superuser:
+            raise PermissionDenied
     form = ExperienceForm(request.POST or None)
     if request.method == 'POST':
         if form.is_valid():
@@ -51,6 +60,8 @@ def create_experience(request):
 
 def update_experience(request, experience_id):
     experience = get_object_or_404(Experience, pk=experience_id)
+    if not request.user.is_superuser:
+            raise PermissionDenied
     form = ExperienceForm(request.POST or None, instance=experience)
     if request.method == 'POST' and form.is_valid():
         form.save()
@@ -66,6 +77,8 @@ def update_experience(request, experience_id):
 
 def delete_experience(request, experience_id):
     experience = get_object_or_404(Experience, pk=experience_id)
+    if not request.user.is_superuser:
+            raise PermissionDenied
     if request.method == 'POST':
         experience.delete()
         messages.success(request, 'Pengalaman berhasil dihapus!')
@@ -92,19 +105,20 @@ def show_skills(request):
     }
     return render(request, "skill.html", context)
 
+@login_required(login_url="/login/") 
 def create_skill(request):
-  form = SkillForm(request.POST or None)
-
-  if request.method == "POST" and form.is_valid():
-    form.save()
-    messages.success(request, "Skill baru berhasil ditambahkan!")
-    return redirect("main:show_skills") 
-  context = {
+    if not request.user.is_superuser:
+            raise PermissionDenied
+    form = SkillForm(request.POST or None)
+    if request.method == "POST" and form.is_valid():
+        form.save()
+        messages.success(request, "Skill baru berhasil ditambahkan!")
+        return redirect("main:show_skills") 
+    context = {
       "name": "Salma Maharani", 
       "form": form,
   }
-
-  return render(request, "skills_form.html", context)
+    return render(request, "skills_form.html", context)
 
 def get_skills_json(request):
     title_query = request.GET.get("title", "").strip()
@@ -113,11 +127,13 @@ def get_skills_json(request):
     if title_query:
         skills = skills.filter(name__icontains=title_query)
 
-    skills_json = serializers.serialize("json", skills)
+    skills_json = serializers.serialize("json", skills, use_natural_foreign_keys=True)
     return HttpResponse(skills_json, content_type="application/json")
 
 def delete_skill(request, skill_id):
    skill = get_object_or_404(Skill, pk=skill_id)
+   if not request.user.is_superuser:
+           raise PermissionDenied
    if request.method == "POST":
         skill.delete()
         messages.success(request, "Skill berhasil dihapus!")
@@ -126,6 +142,8 @@ def delete_skill(request, skill_id):
 
 def update_skill(request, id):
     skill = get_object_or_404(Skill, pk=id)
+    if not request.user.is_superuser:
+            raise PermissionDenied
     form = SkillForm(request.POST or None, instance=skill)
 
     if form.is_valid() and request.method == "POST":
@@ -152,7 +170,10 @@ def show_education(request):
     }
     return render(request, "education.html", context)
 
+@login_required(login_url="/login/") 
 def create_education(request):
+    if not request.user.is_superuser:
+        raise PermissionDenied
     form = EducationForm(request.POST or None)
     if request.method == "POST" and form.is_valid():
         form.save()
@@ -168,6 +189,8 @@ def create_education(request):
 
 def update_education(request, education_id):
     education = get_object_or_404(Education, pk=education_id)
+    if not request.user.is_superuser:
+            raise PermissionDenied
     form = EducationForm(request.POST or None, instance=education)
     if request.method == "POST" and form.is_valid():
         form.save()
@@ -183,6 +206,8 @@ def update_education(request, education_id):
 
 def delete_education(request, education_id):
     education = get_object_or_404(Education, pk=education_id)
+    if not request.user.is_superuser:
+            raise PermissionDenied
     if request.method == "POST":
         education.delete()
         messages.success(request, "Pendidikan berhasil dihapus!")
@@ -198,6 +223,52 @@ def get_education_json(request):
     education_json = serializers.serialize("json", education_list)
     return HttpResponse(education_json, content_type="application/json")
 
+def register(request):
+    form = UserCreationForm(request.POST or None)
+
+    if request.method == "POST" and form.is_valid():
+        form.save()
+        messages.success(request, "Akun berhasil dibuat. Silakan login.")
+        return redirect("main:login")
+
+    context = {
+        "name": "Salma Maharani",
+        "form": form,
+    }
+    return render(request, "register.html", context)
 
 
+def login_user(request):
+    form = AuthenticationForm(request, data=request.POST or None)
 
+    if request.method == "POST" and form.is_valid():
+        user = form.get_user()
+        login(request, user)
+        response = redirect("main:show_main")
+        response.set_cookie('last_login', datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S'))
+        return response
+
+    context = {
+        "name": "Salma Maharani",
+        "form": form,
+    }
+    return render(request, "login.html", context)
+
+def logout_user(request):
+    logout(request)
+    response = redirect("main:show_main")
+    response.delete_cookie('last_login')
+    return response
+
+
+@login_required(login_url="/login/")
+def toggle_star(request, project_id):
+    skill = get_object_or_404(Skill, pk=project_id)
+
+    if request.method == "POST":
+        if request.user in skill.starred_by.all():
+            skill.starred_by.remove(request.user)
+        else:
+            skill.starred_by.add(request.user)
+
+    return redirect("main:show_skill")
